requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
flask-mysql-connector==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.3
pandas==2.2.3
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
setuptools==76.0.0
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.3
wheel==0.45.1
